test(scripts): extract atlas-entry projection helpers and cover them
scripts/build-content-index.mjs defined its Atlas index projection inline, so the
credential-placeholder scrubbing and the entry-to-Atlas field projection behind
the generated atlas-registry.json were untestable without running the whole
build pipeline. Move the pure helpers into scripts/lib/atlas-entry.mjs - matching
the existing scripts/lib convention - and import pickAtlasEntry back.

- scripts/lib/atlas-entry.mjs: scrubAtlasCredentialPlaceholders (with its
  placeholder table) and pickAtlasEntry.
- scripts/build-content-index.mjs: import pickAtlasEntry; drop the local copies
  and the placeholder constant.
- tests: cover placeholder replacement in strings, array/object recursion and
  the pass-through of non-string/non-container values, plus the repoStats
  branches (repoUrl-derived vs stats-only vs absent) and the trustSignals
  present/absent projection. Branch coverage 100% (22/22).

No behavior change; the generated index is identical.
